requirements.txt
pandas
numpy
eikon
progressbar2
psycopg2